Create experiment_set and experiment configs
Modified functionality so that we look in the config files for
"experiment" dictionary in order to populate experiment hparams and
"experiment_set" in order to use config items for ExperimentSet classes,
such as setting the learning rates for all agents.

Added functionality to get these learning rates from the config files.