Save MDP size
Added functionality to save the MDP size as a hyperparameter. In
addition, if no environments exist in the environment directory, then
the MDPs will be created with this size. If MDPs do exist in that
directoy, the sizes of those MDPs will overwrite the value specified in
the config file.