Save actions as hparam
Saved the number of actions as a hyper parameter. Had to remove this
before to get the states working. Now, the number of actions are saved
in the hparam dictionary for ease of use in constructing agents etc.
But, if the number of actions is specified in the configuration file,
it is ignored. The number of actions are solely taken from the
environment.