Fix bug with MDP size
Fixed a bug where the program would crash if there was no mdp_size
specified in the config file and no MDP environments previously created
and existing in the data directory. Now, we set a default MDP size so
that if nothing is specified in the config file and no MDPs exist, we
will create default MDPs of size 8x8.